+++ installer/test/mdaa-installer.snapshot.test.ts
@@ -33,11 +33,6 @@ describe('MdaaInstallerStack Snapshots', () => {
     expect(templateJson.Resources).toMatchSnapshot('resources.json');
   });
 
-  test('Conditions section snapshot', () => {
-    const templateJson = template.toJSON();
-    expect(templateJson.Conditions).toMatchSnapshot('conditions.json');
-  });
-
   test('Rules section snapshot', () => {
     const templateJson = template.toJSON();
     expect(templateJson.Rules).toMatchSnapshot('rules.json');
@@ -50,20 +45,9 @@ describe('MdaaInstallerStack Snapshots', () => {
 
   test('KMS resources snapshot', () => {
     const kmsResources = template.findResources('AWS::KMS::Key');
-    const kmsAliases = template.findResources('AWS::KMS::Alias');
     expect({
       keys: kmsResources,
-      aliases: kmsAliases,
     }).toMatchSnapshot('kms-resources.json');
-  });
-
-  test('S3 resources snapshot', () => {
-    const s3Buckets = template.findResources('AWS::S3::Bucket');
-    const s3BucketPolicies = template.findResources('AWS::S3::BucketPolicy');
-    expect({
-      buckets: s3Buckets,
-      bucketPolicies: s3BucketPolicies,
-    }).toMatchSnapshot('s3-resources.json');
   });
 
   test('IAM resources snapshot', () => {
@@ -78,11 +62,6 @@ describe('MdaaInstallerStack Snapshots', () => {
   test('CodeBuild resources snapshot', () => {
     const codeBuildProjects = template.findResources('AWS::CodeBuild::Project');
     expect(codeBuildProjects).toMatchSnapshot('codebuild-resources.json');
-  });
-
-  test('CodePipeline resources snapshot', () => {
-    const codePipelines = template.findResources('AWS::CodePipeline::Pipeline');
-    expect(codePipelines).toMatchSnapshot('codepipeline-resources.json');
   });
 
   test('Environment variables snapshot', () => {
@@ -105,20 +84,6 @@ describe('MdaaInstallerStack Snapshots', () => {
     expect(codeBuildProject?.Properties?.Source?.BuildSpec).toMatchSnapshot('buildspec.json');
   });
 
-  test('Pipeline stages snapshot', () => {
-    const templateJson = template.toJSON();
-    const pipelines = Object.values(templateJson.Resources).filter(
-      (resource: any) => resource.Type === 'AWS::CodePipeline::Pipeline',
-    ) as any[];
-
-    const pipelineStages = pipelines.map(pipeline => ({
-      name: pipeline.Properties.Name,
-      stages: pipeline.Properties.Stages,
-    }));
-
-    expect(pipelineStages).toMatchSnapshot('pipeline-stages.json');
-  });
-
   test('Parameter groups and labels snapshot', () => {
     const templateJson = template.toJSON();
     const cfnInterface = templateJson.Metadata?.['AWS::CloudFormation::Interface'];
@@ -126,7 +91,6 @@ describe('MdaaInstallerStack Snapshots', () => {
     expect({
       parameterGroups: cfnInterface?.ParameterGroups,
       parameterLabels: cfnInterface?.ParameterLabels,
-      parameterVisibility: cfnInterface?.ParameterVisibility,
     }).toMatchSnapshot('parameter-interface.json');
   });
 
@@ -134,28 +98,14 @@ describe('MdaaInstallerStack Snapshots', () => {
     const templateJson = template.toJSON();
 
     const securityConfigs = {
-      s3Encryption: [],
       kmsKeyRotation: [],
-      sslEnforcement: [],
-      publicAccessBlocks: [],
     } as any;
 
     Object.values(templateJson.Resources).forEach((resource: any) => {
-      if (resource.Type === 'AWS::S3::Bucket') {
-        if (resource.Properties.BucketEncryption) {
-          securityConfigs.s3Encryption.push(resource.Properties.BucketEncryption);
-        }
-        if (resource.Properties.PublicAccessBlockConfiguration) {
-          securityConfigs.publicAccessBlocks.push(resource.Properties.PublicAccessBlockConfiguration);
-        }
-      }
       if (resource.Type === 'AWS::KMS::Key') {
         if (resource.Properties.EnableKeyRotation) {
           securityConfigs.kmsKeyRotation.push(resource.Properties.EnableKeyRotation);
         }
-      }
-      if (resource.Type === 'AWS::S3::BucketPolicy') {
-        securityConfigs.sslEnforcement.push(resource.Properties.PolicyDocument);
       }
     });
 
@@ -173,22 +123,6 @@ describe('MdaaInstallerStack Snapshots', () => {
     });
 
     expect(dependencies).toMatchSnapshot('resource-dependencies.json');
-  });
-
-  test('Conditional resources snapshot', () => {
-    const templateJson = template.toJSON();
-
-    const conditionalResources = {} as any;
-    Object.entries(templateJson.Resources).forEach(([resourceId, resource]: [string, any]) => {
-      if (resource.Condition) {
-        conditionalResources[resourceId] = {
-          type: resource.Type,
-          condition: resource.Condition,
-        };
-      }
-    });
-
-    expect(conditionalResources).toMatchSnapshot('conditional-resources.json');
   });
 });
 
@@ -214,15 +148,7 @@ describe('MdaaInstallerStack with different configurations', () => {
       description: templateJson.Description,
       parameters: Object.keys(templateJson.Parameters || {}),
       resourceCount: Object.keys(templateJson.Resources || {}).length,
-      conditionCount: Object.keys(templateJson.Conditions || {}).length,
       ruleCount: Object.keys(templateJson.Rules || {}).length,
     }).toMatchSnapshot('custom-stack-summary.json');
-  });
-
-  test('Repository source enum values snapshot', () => {
-    expect({
-      GITHUB: MdaaInstaller.RepositorySources.GITHUB,
-      S3: MdaaInstaller.RepositorySources.S3,
-    }).toMatchSnapshot('repository-sources.json');
   });
 });